fix(Windows): I hate verbose windows builds.
Untested.

+++ hammer/VGuiWnd.h
@@ -6,7 +6,7 @@
 namespace vgui
 {
 	class EditablePanel;
-	typedef unsigned long HCursor;
+	typedef unsigned int HCursor;
 }
 
 class CVGuiWnd
