+++ public/vgui/ISurfaceV30.h
@@ -43,9 +43,9 @@ namespace vgui
 	class Point;
 
 	// handles
-	typedef unsigned long HCursor;
-	typedef unsigned long HTexture;
-	typedef unsigned long HFont;
+	typedef uint32 HCursor;
+	typedef uint32 HTexture;
+	typedef uint32 HFont;
 }
